refactor: split jsonrpc params and openapi fragments #261
